+++ src/lib/services/account.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { z } from "zod";
 
-import { zProjectInfo, zPublicAccountInfo } from "lib/types";
+import { zProjectInfo, zPublicAccountInfo, type Addr } from "lib/types";
 
 const zIcns = z.object({
   names: z.array(z.string()),
@@ -24,8 +24,37 @@ export type AccountInfo = z.infer<typeof zAccountInfo>;
 
 export const getAccountInfo = async (
   endpoint: string,
-  address: string
+  address: Addr
 ): Promise<AccountInfo> =>
   axios
     .get(`${endpoint}/${encodeURIComponent(address)}/info`)
     .then((res) => zAccountInfo.parse(res.data));
+
+const zAccountTableCounts = z
+  .object({
+    code: z.number().nullish(),
+    contract_by_admin: z.number().nullish(),
+    instantiated: z.number().nullish(),
+    proposal: z.number().nullish(),
+    tx: z.number().nullish(),
+  })
+  .transform((tableCount) => ({
+    code: tableCount.code,
+    contractByAdmin: tableCount.contract_by_admin,
+    instantiated: tableCount.instantiated,
+    proposal: tableCount.proposal,
+    tx: tableCount.tx,
+  }));
+
+export type AccountTableCounts = z.infer<typeof zAccountTableCounts>;
+
+export const getAccountTableCounts = async (
+  endpoint: string,
+  address: string,
+  isWasm: boolean
+): Promise<AccountTableCounts> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(address)}/table-counts`, {
+      params: { is_wasm: isWasm },
+    })
+    .then((res) => zAccountTableCounts.parse(res.data));